+++ R/method_cc.R
@@ -72,13 +72,18 @@ cc <- function(mdata, base.method = getOption("utiml.base.method", "SVM"),
 
   # Create models
   basedata <- mdata$dataset[mdata$attributesIndexes]
-  labeldata <- mdata$dataset[mdata$labels$index][chain]
+  labeldata <- as.data.frame(mdata$dataset[mdata$labels$index][chain])
+  for (i in seq(ncol(labeldata))) {
+    labeldata[, i] <- factor(labeldata[, i], levels=c(0, 1))
+  }
+
   chain.order <- utiml_rename(seq(mdata$measures$num.labels), chain)
   ccmodel$models <- utiml_lapply(chain.order, function(lidx) {
-    data <- cbind(basedata, labeldata[seq(lidx)])
-    dataset <- utiml_prepare_data(data, "mldCC", mdata$name, "cc", base.method,
-                               chain.order = lidx)
-    utiml_create_model(dataset, ...)
+    utiml_create_model(
+      utiml_prepare_data(
+        cbind(basedata, labeldata[seq(lidx)]),
+        "mldCC", mdata$name, "cc", base.method, chain.order = lidx
+      ), ...)
   }, cores, seed)
 
   utiml_restore_seed()
@@ -134,10 +139,9 @@ predict.CCmodel <- function(object, newdata,
   newdata <- list(utiml_newdata(newdata))
   predictions <- list()
   for (label in object$chain) {
-    predictions[[label]] <- utiml_predict_binary_model(object$models[[label]],
-                                                       do.call(cbind, newdata),
-                                                       ...)
-    newdata[[label]] <- predictions[[label]]$bipartition
+    predictions[[label]] <- utiml_predict_binary_model(
+      object$models[[label]], do.call(cbind, newdata), ...)
+    newdata[[label]] <- factor(predictions[[label]]$bipartition, levels=c(0, 1))
   }
 
   utiml_restore_seed()