+++ R/method_mbr.R
@@ -104,7 +104,6 @@ mbr <- function(mdata, base.method = getOption("utiml.base.method", "SVM"),
   else {
     kf <- create_kfold_partition(mdata, folds, "iterative")
     base.preds <- do.call(rbind, lapply(seq(folds), function(f) {
-      #TODO label with few positive examples
       dataset <- partition_fold(kf, f)
       classifier <- br(dataset$train, base.method, ..., cores=cores, seed=seed)
       params <- list(object = classifier, newdata = dataset$test,
@@ -114,19 +113,28 @@ mbr <- function(mdata, base.method = getOption("utiml.base.method", "SVM"),
     base.preds <- base.preds[rownames(mdata$dataset), ]
   }
 
+  base.preds <- as.data.frame(base.preds)
+  for (i in seq(ncol(base.preds))) {
+    base.preds[, i] <- factor(base.preds[, i], levels=c(0, 1))
+  }
+
   # 2 Iteration - Meta level -------------------------------------------------
-  corr <- mbrmodel$correlation <- utiml_labels_correlation(mdata)
+  corr <- abs(stats::cor(mdata$dataset[mdata$labels$index]))
+  mbrmodel$correlation <- corr
   labels <- utiml_rename(mbrmodel$labels)
   mbrmodel$models <- utiml_lapply(labels, function (label) {
     nmcol <- colnames(corr)[corr[label, ] >= phi]
     new.data <- base.preds[, nmcol, drop = FALSE]
-    colnames(new.data) <- paste("extra", nmcol, sep = ".")
+    if (ncol(new.data) > 0) {
+      colnames(new.data) <- paste("extra", nmcol, sep = ".")
+    }
 
-    brdata  <- utiml_create_binary_data(mdata, label, new.data)
-    dataset <- utiml_prepare_data(brdata, "mldMBR", mdata$name, "mbr",
-                                  base.method, new.features = nmcol)
-
-    utiml_create_model(dataset, ...)
+    utiml_create_model(
+      utiml_prepare_data(
+        utiml_create_binary_data(mdata, label, new.data),
+        "mldMBR", mdata$name, "mbr", base.method, new.features = nmcol
+      ), ...
+    )
   }, cores, seed)
 
   utiml_restore_seed()
@@ -186,6 +194,10 @@ predict.MBRmodel <- function(object, newdata,
   base.preds <- as.bipartition(predict.BRmodel(object$basemodel, newdata,
                                                probability=FALSE, ...,
                                                cores=cores, seed=seed))
+  base.preds <- as.data.frame(base.preds)
+  for (i in seq(ncol(base.preds))) {
+    base.preds[,i] <- factor(base.preds[,i], levels=c(0, 1))
+  }
 
   # 2 Iteration - Meta level -------------------------------------------------
   corr <- object$correlation
@@ -193,7 +205,10 @@ predict.MBRmodel <- function(object, newdata,
   predictions <- utiml_lapply(labels, function(labelname) {
     nmcol <- colnames(corr)[corr[labelname, ] >= object$phi]
     extra.col <- base.preds[, nmcol, drop = FALSE]
-    colnames(extra.col) <- paste("extra", nmcol, sep = ".")
+    if (ncol(extra.col) > 0) {
+      colnames(extra.col) <- paste("extra", nmcol, sep = ".")
+    }
+
     utiml_predict_binary_model(object$models[[labelname]],
                                cbind(newdata, extra.col), ...)
   }, cores, seed)
@@ -231,18 +246,23 @@ predict.MBRmodel <- function(object, newdata,
 #' ## result[4, -4]
 utiml_labels_correlation <- function(mdata) {
   label.names <- rownames(mdata$labels)
-  classes <- mdata$dataset[, mdata$labels$index]
+  classes <- lapply(mdata$labels$index, function (col) {
+    factor(mdata$dataset[, col], levels=c("0", "1"))
+  })
+
   q <- length(label.names)
   cor <- matrix(nrow = q, ncol = q, dimnames = list(label.names, label.names))
-  for (i in 1:q) {
-    for (j in i:q) {
-      confmat <- table(classes[, c(i, j)])
+
+  for (i in seq(1, q)) {
+    for (j in seq(i, q)) {
+      confmat <- table(classes[c(i, j)])
       A <- as.numeric(confmat["1", "1"])
       B <- as.numeric(confmat["1", "0"])
       C <- as.numeric(confmat["0", "1"])
       D <- as.numeric(confmat["0", "0"])
-      cor[i, j] <- abs((A * D - B * C) / sqrt(as.numeric(A + B) * (C + D) *
-                                                (A + C) * (B + D)))
+      value1 <- A * D - B * C
+      value2 <- sqrt(as.numeric(A + B) * (C + D) * (A + C) * (B + D))
+      cor[i, j] <- ifelse(value1 == 0 & value2 == 0, -Inf, abs(value1 / value2))
       cor[j, i] <- cor[i, j]
     }
   }