+++ R/method_rdbr.R
@@ -143,21 +143,33 @@ predict.RDBRmodel <- function(object, newdata, estimative = NULL,
                                   cores=cores, seed=seed)
   }
 
+  if ('mlresult' %in% class(estimative)) {
+    estimative <- as.bipartition(estimative)
+  }
+
+  estimative <- as.data.frame(estimative)
+  for (i in seq(ncol(estimative))) {
+    estimative[,i] <- factor(estimative[,i], levels=c(0, 1))
+  }
+
   labels <- names(object$models)
   modelsindex <- utiml_rename(seq(labels), labels)
   if (batch.mode) {
     for (i in seq(max.iterations)) {
+      old.estimative <- estimative
       predictions <- utiml_lapply(modelsindex, function(li) {
         utiml_predict_binary_model(object$models[[li]],
                                    cbind(newdata, estimative[, -li]), ...)
       }, cores, seed)
 
-      new.estimative <- do.call(cbind, lapply(predictions,
-                                              function(lbl) lbl$bipartition))
-      if (all(new.estimative == estimative)) {
+      for (j in seq(predictions)) {
+        classes <- predictions[[j]]$bipartition
+        estimative[, j] <- factor(classes, levels=c(0, 1))
+      }
+
+      if (all(old.estimative == estimative)) {
         break
       }
-      estimative <- new.estimative
     }
   }
   else {
@@ -166,12 +178,13 @@ predict.RDBRmodel <- function(object, newdata, estimative = NULL,
       predictions <- list()
 
       # the labels needs to be shuffled in each iteraction
-      for (li in modelsindex) {
+      for (li in sample(modelsindex)) {
         predictions[[li]] <- utiml_predict_binary_model(object$models[[li]],
                                               cbind(newdata, estimative[, -li]),
                                               ...)
-        estimative[, li] <- predictions[[li]]$bipartition
+        estimative[, li] <- factor(predictions[[li]]$bipartition, levels=c(0,1))
       }
+
       names(predictions) <- labels
       if (all(old.estimative == estimative)) {
         break