+++ NEWS.md
@@ -1,6 +1,39 @@
 # Changelog
 
-## utiml 0.1.0.91 (Ankle)
+## utiml 0.1.1
+New multi-label transformation methods including pairwise and multiclass 
+approaches. Some fixes from previous version.
+
+### Major changes
+* lcard threshold calibration
+* Use categorical attributes in multilabel datasets and methods
+* LIFT multi-label classification method
+* RPC multi-label classification method
+* CRL multi-label classification method
+* LP multi-label classification method
+* RAkEL multi-label classification method
+* BASELINE multi-label classification method
+* PPT multi-label classification method
+* PS multi-label classification method
+* EPS multi-label classification method
+* HOMER multi-label classification method
+
+### Minor changes
+* Add Empty Model as base method to fix training labels with few examples
+* `multilabel_confusion_matrix` accepts a data.frame or matrix with the predicitons
+* Change EBR and ECC to use threshold calibration
+* Include empty.prediction configuration to enable/disable empty predictions
+
+### Bug fixes
+* Majority Ensemble Predictions Votes
+* Majority Ensemble Predictions Probability
+* Base method not found message error
+* Base method support any attribute names
+* Normalize data ignore attributes with a single value
+* MBR support labels without positive examples
+* Fix average precision and coverage measures to support instances without labels
+
+## utiml 0.1.0
 
 First release of **utiml**: